Add the --open flag to all nmap scans
The script (before the POSIX changes) grepped for the open ports in every scan result before doing some other automated scan/task with them.

The --open nmap flag only outputs the open ports, making irrelevant to grep for the open ports as they are they only ones being showed. The non-open ports appear in a summary like the following: `Not shown: 949 closed ports, 47 filtered ports`

It shall be noted that the filtered ports won't appear when --open is specified. To maybe allow the user to see the full list of ports, nmapAutomator.sh could receive a flag for this purpose (like --show-all-ports), which basically don't pass the --open flag to nmap.